+++ src/util.test.ts
@@ -18,44 +18,6 @@ describe('util', () => {
         },
       ])
     })
-
-    test('should generate pipeline from omit with dotted fields', () => {
-      const pipeline = generatePipelineFromOmit([
-        'documents.agenda.parsedText',
-        'documents.agenda.contentType',
-        'createdAt',
-      ])
-      assert.deepEqual(pipeline, [
-        {
-          $unset: [
-            'fullDocument.documents.agenda.parsedText',
-            'updateDescription.updatedFields.documents.agenda.parsedText',
-            'fullDocument.documents.agenda.contentType',
-            'updateDescription.updatedFields.documents.agenda.contentType',
-            'fullDocument.createdAt',
-            'updateDescription.updatedFields.createdAt',
-          ],
-        },
-        {
-          $set: {
-            'updateDescription.updatedFields': {
-              $arrayToObject: {
-                $filter: {
-                  input: { $objectToArray: '$updateDescription.updatedFields' },
-                  cond: {
-                    $regexMatch: {
-                      input: '$$this.k',
-                      regex:
-                        '^(?!documents\\.agenda\\.parsedText|documents\\.agenda\\.contentType)',
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      ])
-    })
   })
   describe('removeUnusedFields', () => {
     test('should remove all unneeded fields', () => {